+++ include/LIEF/MachO/Header.hpp
@@ -195,14 +195,14 @@ class LIEF_API Header : public Object {
 
   /// True if the binary is 32-bit
   bool is_32bit() const {
-    return magic_ == MACHO_TYPES::MH_MAGIC ||
-           magic_ == MACHO_TYPES::MH_CIGAM;
+    return magic_ == MACHO_TYPES::MAGIC ||
+           magic_ == MACHO_TYPES::CIGAM;
   }
 
   /// True if the binary is 64-bit
   bool is_64bit() const {
-    return magic_ == MACHO_TYPES::MH_MAGIC_64 ||
-           magic_ == MACHO_TYPES::MH_CIGAM_64;
+    return magic_ == MACHO_TYPES::MAGIC_64 ||
+           magic_ == MACHO_TYPES::CIGAM_64;
   }
 
   void remove(FLAGS flag);